server-experiment/requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
Flask-PyMongo==2.2.0
flask-cors==3.0.10
requests==2.21.0
pandas==2.1.4
nanoid==2.0.0
watchdog==6.0.0
sqlalchemy>=2.0
psycopg2-binary>=2.9